Refactor: Extract submission photo upload flow to PhotoUploadDelegate
- Extracted `uploadSubmitPhotos` from `UploadManager` into a new dedicated `PhotoUploadDelegate` class.
- Modified `UploadManager` to delegate to `PhotoUploadDelegate`, reducing its responsibilities.
- Updated `ServiceModule` to inject the new dependency and pass it to `UploadManager`.
- Updated `UploadManagerTest` to maintain full test coverage for the delegated behavior.
