+++ app/src/main/java/org/ole/planet/myplanet/services/UploadManager.kt
@@ -32,6 +32,7 @@ import org.ole.planet.myplanet.repository.UserRepository
 import org.ole.planet.myplanet.services.upload.UploadConfigs
 import org.ole.planet.myplanet.services.upload.UploadCoordinator
 import org.ole.planet.myplanet.services.upload.UploadResult
+import org.ole.planet.myplanet.services.upload.delegate.PhotoUploadDelegate
 import org.ole.planet.myplanet.utils.FileUtils
 import org.ole.planet.myplanet.utils.JsonUtils.getString
 import org.ole.planet.myplanet.utils.NetworkUtils
@@ -67,7 +68,8 @@ class UploadManager @Inject constructor(
     private val apiInterface: ApiInterface,
     private val activitiesRepository: org.ole.planet.myplanet.repository.ActivitiesRepository,
     private val dispatcherProvider: org.ole.planet.myplanet.utils.DispatcherProvider,
-    @ApplicationScope private val scope: CoroutineScope
+    @ApplicationScope private val scope: CoroutineScope,
+    private val photoUploadDelegate: PhotoUploadDelegate
 ) : FileUploader(apiInterface, scope) {
 
     private suspend fun uploadNewsActivities() {
@@ -202,54 +204,7 @@ class UploadManager @Inject constructor(
     }
 
     suspend fun uploadSubmitPhotos(listener: OnSuccessListener?) {
-        val photosToUpload = submissionsRepository.getUnuploadedPhotos()
-
-        if (photosToUpload.isEmpty()) {
-            notifyListener(listener, "No photos to upload")
-            return
-        }
-
-        withContext(dispatcherProvider.io) {
-            data class UploadedPhotoInfo(val photoId: String, val rev: String, val id: String)
-
-            photosToUpload.chunked(BATCH_SIZE).forEach { batch ->
-                val successfulUploads = mutableListOf<UploadedPhotoInfo>()
-
-                batch.forEach { (photoId, serialized) ->
-                    try {
-                        val `object` = apiInterface.postDoc(
-                            UrlUtils.header, "application/json",
-                            "${UrlUtils.getUrl()}/submissions", serialized
-                        ).body()
-
-                        if (`object` != null) {
-                            val rev = getString("rev", `object`)
-                            val id = getString("id", `object`)
-
-                            submissionsRepository.markPhotoUploaded(photoId, rev, id)
-
-                            if (listener != null && photoId != null) {
-                                successfulUploads.add(UploadedPhotoInfo(photoId, rev, id))
-                            }
-                        }
-                    } catch (e: Exception) {
-                        Log.e(TAG, "Exception in UploadManager", e)
-                    }
-                }
-
-                if (listener != null && successfulUploads.isNotEmpty()) {
-                    val photoIds = successfulUploads.map { it.photoId }.toTypedArray()
-                    val photos = submissionsRepository.getPhotosByIds(photoIds)
-
-                    photos.forEach { photo ->
-                        val uploadInfo = successfulUploads.find { it.photoId == photo.id }
-                        if (uploadInfo != null) {
-                            uploadAttachment(uploadInfo.id, uploadInfo.rev, photo, listener)
-                        }
-                    }
-                }
-            }
-        }
+        photoUploadDelegate.uploadSubmitPhotos(this, listener, ::notifyListener)
     }
 
     suspend fun uploadResource(listener: OnSuccessListener?) {
